refactor(workflow): move ORCA stage helpers v1_monitor -> orca_progress (#224 Phase 1)
